requirements.txt
pillow
pdf2image
PyPDF2
requests
striprtf
python-docx
torch
sentence-transformers
scipy
gensim
en
